AuctionService use authentication

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -60,13 +61,13 @@ public  class AuctionsController : ControllerBase
         return mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
     {
         var auction = mapper.Map<Auction>(auctionDto);
 
-        //TODO: add current user as seller
-        auction.Seller = "test";
+        auction.Seller = User?.Identity?.Name ?? "Guest";
 
         context.Auctions.Add(auction);
 
@@ -90,10 +91,9 @@ public  class AuctionsController : ControllerBase
                 .Include(x => x.Item)
                 .FirstOrDefaultAsync(x => x.Id == id);
 
-        if (auction == null)
-        {
-            return NotFound();
-        }
+        if (auction == null) return NotFound();
+
+        if(auction.Seller != User.Identity?.Name) return Forbid();
         
         auction.Item.Make = auctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = auctionDto.Model ?? auction.Item.Model;
@@ -116,6 +116,8 @@ public  class AuctionsController : ControllerBase
                 .FirstOrDefaultAsync(x => x.Id == id);
 
         if (auction == null) return NotFound();
+
+        if(auction.Seller != User.Identity?.Name) return Forbid();
 
         context.Auctions.Remove(auction);
